+++ NEWS.md
@@ -52,6 +52,8 @@
 
 - `variationPlot()` and `timeVariation()` will no longer plot a caption referencing an irrelevant confidence interval if `ci = FALSE`.
 
+- Restored functionality of `polarPlot(upper=)`.
+
 # openair 3.1.0
 
 ## Dependency Changes